+++ test/mocks/network/mocks.h
@@ -314,6 +314,7 @@ public:
   MOCK_CONST_METHOD0(name, const std::string&());
   MOCK_METHOD0(udpListenerFactory, const Network::ActiveUdpListenerFactory*());
   MOCK_METHOD0(connectionBalancer, ConnectionBalancer&());
+  MOCK_METHOD0(openConnections, ResourceLimit&());
 
   envoy::api::v2::core::TrafficDirection direction() const override {
     return envoy::api::v2::core::TrafficDirection::UNSPECIFIED;