+++ test/mocks/runtime/mocks.cc
@@ -1,11 +1,14 @@
 #include "mocks.h"
 
+#include "common/common/empty_string.h"
+
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 
 using testing::_;
 using testing::Return;
 using testing::ReturnArg;
+using testing::ReturnRef;
 
 namespace Envoy {
 namespace Runtime {
@@ -18,6 +21,7 @@ MockSnapshot::MockSnapshot() {
   ON_CALL(*this, getInteger(_, _)).WillByDefault(ReturnArg<1>());
   ON_CALL(*this, getDouble(_, _)).WillByDefault(ReturnArg<1>());
   ON_CALL(*this, getBoolean(_, _)).WillByDefault(ReturnArg<1>());
+  ON_CALL(*this, get(_)).WillByDefault(ReturnRef(EMPTY_STRING));
 }
 
 MockSnapshot::~MockSnapshot() = default;