+++ docs/root/intro/version_history.rst
@@ -3,6 +3,7 @@ Version history
 
 1.12.5 (Pending)
 ================
+* listener: add runtime support for `per-listener limits <config_listeners_runtime>` on active/accepted connections.
 
 1.12.4 (June 8, 2020)
 =====================